extracted modelspec subclasses
