+++ Nictoarch.Modelling.Core/ISourceFactory.cs
@@ -1,14 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Jsonata.Net.Native.Json;
-using Nictoarch.Modelling.Core.Elements;
-using Nictoarch.Modelling.Core.Yaml;
-using YamlDotNet.Serialization;
-using YamlDotNet.Serialization.BufferedDeserialization;
+using Nictoarch.Modelling.Core.Spec;
 using YamlDotNet.Serialization.BufferedDeserialization.TypeDiscriminators;
 
 namespace Nictoarch.Modelling.Core
@@ -19,21 +14,21 @@ namespace Nictoarch.Modelling.Core
         IEnumerable<ITypeDiscriminator> GetYamlTypeDiscriminators();
     }
 
-    public interface ISource: IAsyncDisposable
+    public interface ISource : IAsyncDisposable
     {
 
     }
 
-    public interface ISource<TExtractConfig>: ISource
-        where TExtractConfig : ModelSpec.ExtractConfigBase
+    public interface ISource<TExtractConfig> : ISource
+        where TExtractConfig : ExtractConfigBase
     {
         Task<JToken> Extract(TExtractConfig extractConfig, CancellationToken cancellationToken);
     }
 
     public interface ISourceFactory<TSourceConfig, TSource, TExtractConfig> : ISourceFactory
-        where TSourceConfig : ModelSpec.SourceConfigBase
-        where TExtractConfig: ModelSpec.ExtractConfigBase
-        where TSource: ISource<TExtractConfig>
+        where TSourceConfig : SourceConfigBase
+        where TExtractConfig : ExtractConfigBase
+        where TSource : ISource<TExtractConfig>
     {
         Task<ISource> GetSource(TSourceConfig sourceConfig, CancellationToken cancellationToken);
     }
