+++ Nictoarch.Modelling.Core/Spec/ExtractConfigBase.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Nictoarch.Modelling.Core.Spec
+{
+    public abstract class ExtractConfigBase
+    {
+
+    }
+}
