+++ Nictoarch.Modelling.Json/JsonExtractConfig.cs
@@ -4,11 +4,11 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Nictoarch.Modelling.Core;
+using Nictoarch.Modelling.Core.Spec;
 
 namespace Nictoarch.Modelling.Json
 {
-    internal class JsonExtractConfig: ModelSpec.ExtractConfigBase
+    internal class JsonExtractConfig: ExtractConfigBase
     {
         public enum ESourceTransform
         {
